Move createContinuousRange{,Or}UntypedTarget to UntypedTarget.ts.
