+++ .vscode/settings.json
@@ -25,5 +25,9 @@
   "files.eol": "\n",
   "typescript.enablePromptUseWorkspaceTsdk": true,
   "typescript.tsdk": "node_modules/typescript/lib",
-  "eslint.workingDirectories": [{ "pattern": "packages/*/" }]
+  "eslint.workingDirectories": [
+    {
+      "pattern": "packages/*/"
+    }
+  ]
 }
