+++ packages/cursorless-engine/src/processTargets/targetUtil/createContinuousRange.ts
@@ -1,7 +1,5 @@
 import { Position, Range } from "@cursorless/common";
 import type { Target } from "../../typings/target.types";
-import { UntypedTarget } from "../targets";
-import { isSameType } from "../../util/typeUtils";
 
 export function createContinuousRange(
   startTarget: Target,
@@ -47,55 +45,4 @@ export function createContinuousLineRange(
   return new Range(start, end);
 }
 
-export function createContinuousRangeUntypedTarget(
-  isReversed: boolean,
-  startTarget: Target,
-  endTarget: Target,
-  includeStart: boolean,
-  includeEnd: boolean,
-): UntypedTarget {
-  return new UntypedTarget({
-    editor: startTarget.editor,
-    isReversed,
-    hasExplicitRange: true,
-    contentRange: createContinuousRange(
-      startTarget,
-      endTarget,
-      includeStart,
-      includeEnd,
-    ),
-    isToken: startTarget.isToken && endTarget.isToken,
-  });
-}
 
-export function createContinuousRangeOrUntypedTarget(
-  isReversed: boolean,
-  startTarget: Target,
-  cloneParameters: any,
-  endTarget: Target,
-  includeStart: boolean,
-  includeEnd: boolean,
-): Target {
-  if (isSameType(startTarget, endTarget)) {
-    const constructor = Object.getPrototypeOf(startTarget).constructor;
-
-    return new constructor({
-      ...cloneParameters,
-      isReversed,
-      contentRange: createContinuousRange(
-        startTarget,
-        endTarget,
-        includeStart,
-        includeEnd,
-      ),
-    });
-  }
-
-  return createContinuousRangeUntypedTarget(
-    isReversed,
-    startTarget,
-    endTarget,
-    includeStart,
-    includeEnd,
-  );
-}
